Add slew rate limit, add ramp function
Add slew rate limit to the Teensy.ino code. Expressed in micros./step. Step size is expressed in units of LSB. The slew rate can be different for each channel, but the step size is the same for all channels. When the slew rate is nonzero, it is no longer possible to change all channels at once; each must be set individually.
Also add ramp function to the Python terminal code. The function steps from a start to an end voltage and takes an ADC reading for each step.