Add ramp with dither, improve FFT
Add a ramp function that uses dither (not tested yet)
Add a GetConversionDataFast function in the AD7606 library which can get the conversion results from channel 0 much more quickly. This function does not process the data, and simply adds it to an array passed by reference.
Add an FFT function to the Python daq module (originally, the FFT function was only in the Python terminal program)
Fix a bug in the AD7606 library
Add comments in various areas
The FFT functionality is slow; not sure why. Also there are some issues with transferring data from the Teensy to the host computer - sometimes not all of the data gets transferred.