Fix dither ramp plotting
Fix plotting for dither ramp.
Change daq.setup so that the default port is "COM7" if unspecified.
Flush serial USB input and output before requesting data.
If fast-sample data is missing some data, re-request the data up to a maximum number of retries until the transfer is successful.
Add lots of comments.