Convert voltage range and step size
The DAC has discrete output voltages and a limited range. For this commit I added a function to convert a voltage range (consisting of a start voltage, finish voltage, and number of steps) into a voltage range that uses only allowable DAC outputs.
I also added a function to convert a voltage value into an allowable DAC output voltage.